+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v 1.11.0 (2024-08-02)
+Changes in this release:
+- Update command to generate module documentation.
+- Use the README.md file to control what is included in the documentation.
+
 # v 1.10.0 (2024-06-12)
 Changes in this release:
 - Use system pip index to install modules